Add static_cast to avoid clang report syntax error.

+++ src/primihub/operator/aby3_operator.cc
@@ -98,7 +98,8 @@ void MPCOperator::createShares(si64Matrix &sharedMatrix) {
   enc.remoteIntMatrix(runtime, sharedMatrix).get();
 }
 si64Matrix MPCOperator::createSharesByShape(const i64Matrix &val) {
-  std::array<u64, 2> size{val.rows(), val.cols()};
+  std::array<u64, 2> size{static_cast<unsigned long long>(val.rows()),
+                          static_cast<unsigned long long>(val.cols())};
   mNext.asyncSendCopy(size);
   mPrev.asyncSendCopy(size);
   si64Matrix dest(size[0], size[1]);
@@ -122,7 +123,8 @@ si64Matrix MPCOperator::createSharesByShape(u64 pIdx) {
 
 // only support val is column vector
 sbMatrix MPCOperator::createBinSharesByShape(i64Matrix &val, u64 bitCount) {
-  std::array<u64, 2> size{val.rows(), bitCount};
+  std::array<u64, 2> size{static_cast<unsigned long long>(val.rows()),
+                          static_cast<unsigned long long>(bitCount)};
   mNext.asyncSendCopy(size);
   mPrev.asyncSendCopy(size);
   sbMatrix dest(size[0], size[1]);
